Revert "refactor: move db to separate package"
This reverts commit 5a241117bcd29a3000e4aa060b055a58307c0747.
