+++ apps/dashboard/drizzle.config.ts
@@ -2,7 +2,7 @@ import 'dotenv/config';
 import type { Config } from 'drizzle-kit';
 
 export default {
-  schema: './src/schema.ts',
+  schema: './src/lib/db/schema.ts',
   out: './drizzle',
   driver: 'turso',
   dbCredentials: {
