+++ apps/dashboard/package.json
@@ -8,22 +8,26 @@
     "start": "next start",
     "lint": "next lint",
     "typecheck": "tsc --noEmit",
-    "test": "vitest"
+    "test": "vitest",
+    "db:generate": "drizzle-kit generate:sqlite",
+    "db:migrate": "tsx drizzle/migrate.ts",
+    "db:studio": "drizzle-kit studio"
   },
   "dependencies": {
     "@dnd-kit/core": "^6.1.0",
     "@dnd-kit/modifiers": "^7.0.0",
     "@dnd-kit/sortable": "^8.0.0",
     "@dnd-kit/utilities": "^3.2.2",
+    "@libsql/client": "0.4.0-pre.5",
     "@lucia-auth/adapter-drizzle": "1.0.0-beta.0",
-    "@ogstudio/db": "workspace:*",
     "@resvg/resvg-wasm": "^2.6.0",
     "@tanstack/react-query": "^5.15.5",
+    "arctic": "^0.10.2",
+    "drizzle-orm": "^0.29.2",
+    "lucia": "3.0.0-beta.13",
     "@vercel/analytics": "^1.1.1",
     "@vercel/speed-insights": "^1.0.2",
-    "arctic": "^0.10.2",
     "clsx": "^2.1.0",
-    "lucia": "3.0.0-beta.13",
     "next": "14.0.4",
     "oslo": "^0.26.2",
     "react": "^18",
@@ -43,6 +47,7 @@
     "@types/react-dom": "^18",
     "autoprefixer": "^10.0.1",
     "dotenv": "^16.3.1",
+    "drizzle-kit": "^0.20.9",
     "happy-dom": "^12.10.3",
     "postcss": "^8",
     "tailwindcss": "^3.3.0",
