--- packages/db/.env.example
@@ -1,3 +1,0 @@
-DATABASE_URL=
-DATABASE_AUTH_TOKEN=
-
