--- packages/db/src/index.ts
@@ -1,3 +1,0 @@
-export * from './db'
-export * from './schema'
-export * from 'drizzle-orm'
