--- packages/db/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "extends": "@ogstudio/typescript/library.json",
-  "include": [
-    "**/*.ts",
-  ],
-  "exclude": [
-    "node_modules"
-  ]
-}
